requirements.txt
asgiref==3.3.1
Django==3.1.7
pytz==2021.1
psycopg2>=2.8
redis>=3.5
celery>=5.0
requests==2.24.0
djangorestframework==3.13.1
django-celery-beat==2.2.1